Sentiment-For-Casper/requirements.txt
# =============================================================================
# Sentiment For-Casper Project - Requirements
# =============================================================================
# This file lists all Python packages needed to run the project:
# - Data extraction, sentiment analysis, correlation, forecasting, dashboard.
#
# To install all dependencies, run:
#   pip install -r requirements.txt
#
# Python 3.10+ is recommended.
# =============================================================================

# --- Core ---
python-dotenv==1.0.0         # Load environment variables from .env files
pymongo==4.6.0               # MongoDB client for Python
certifi==2023.7.22           # SSL certificates for secure MongoDB connections

# --- Data Processing ---
numpy==1.24.4                # Numerical computing
pandas==2.0.3                # Data analysis and manipulation
python-dateutil==2.8.2       # Date parsing and utilities

# --- Financial Data ---
yfinance==0.2.18             # Download historical market data from Yahoo Finance

# --- Machine Learning & NLP ---
torch==2.0.1                 # PyTorch deep learning framework
scikit-learn==1.3.0          # Machine learning algorithms and utilities
scipy==1.11.1                # Scientific computing (required by scikit-learn)
transformers==4.31.0         # HuggingFace Transformers for NLP models (e.g., FinBERT)
protobuf==4.23.4             # Protocol buffers (required by transformers/torch)
sentencepiece==0.1.99        # Tokenizer for some transformer models
tokenizers==0.13.3           # Fast tokenizers for NLP

# --- Visualization ---
matplotlib==3.7.2            # Plotting library
seaborn==0.12.2              # Statistical data visualization

# --- Web & API ---
flask==2.3.2                 # Web framework for the dashboard
flask-cors==4.0.0            # CORS support for Flask APIs
requests==2.31.0             # HTTP requests library
gnews==0.2.7                 # Google News API wrapper
urllib3==2.0.4               # HTTP library (dependency for requests)

# --- Time & Utilities ---
pytz==2023.3                 # Timezone support
tzlocal==5.0.1               # Local timezone utilities
typing-extensions==4.7.1     # Backport of Python typing features

# --- Optional (install if needed) ---
# jupyter      # For Jupyter notebooks
# plotly       # For interactive plots
# PyYAML       # For YAML

# =============================================================================
# END OF REQUIREMENTS
#